removed Helpers::fixStack()
xdebug_get_function_stack() is available only if xdebug.mode=develop is set.
Inside a shutdown handler, it returns the complete callstack only when called from the CLI.
When used in a browser, the callstack is printed out and the function then does not return it.
The callstack output can be disabled using html_errors=0, but the function still won't return it.